wallet/signing: hash writers and serialization moved to seperate file
